use ValidationException in Create/Update Handlers. use NotFoundException in Delete Handlers. user BaseCommandResponse to create LeaveRequest.

+++ HR_Management.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using HR_Management.Application.Features.LeaveAllocations.Requests.Commands;
+﻿using HR_Management.Application.Exceptions;
+using HR_Management.Application.Features.LeaveAllocations.Requests.Commands;
 using HR_Management.Application.Persistence.Contracts;
+using HR_Management.Domain;
 using MediatR;
 
 namespace HR_Management.Application.Features.LeaveAllocations.Handlers.Commands;
@@ -16,6 +18,10 @@ public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAl
     public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
     {
         var leaveAllocation = await _leaveAllocationRepo.Get(request.Id);
+
+        if (leaveAllocation == null)
+            throw new NotFoundException(nameof(LeaveAllocation), request.Id);
+
         await _leaveAllocationRepo.Delete(leaveAllocation);
         return Unit.Value;
     }
